Moved test helper function to testResources and added a new one to simplify Strategy tests
